requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=1.8.0+cpu
torchvision>=0.9.0
fastapi>=0.65.0
uvicorn>=0.13.0
opencv-python>=4.5.0
pillow>=8.2.0
pyyaml>=5.4.0
python-multipart>=0.0.5
loguru